+++ .gitignore
@@ -10,7 +10,7 @@ MANIFEST
 /env
 # complied cffi output
 libqtile/_ffi_pango.py
-libqtile/core/_ffi_xcursors.py
+libqtile/backend/x11/_ffi_xcursors.py
 # Files generated by setup.py
 qtile.egg-info/
 .eggs
