+++ libqtile/backend/x11/xcursors.py
@@ -1,9 +1,9 @@
-from ..log_utils import logger
+from libqtile.log_utils import logger
 
 try:
     from ._ffi_xcursors import ffi
 except ImportError:
-    raise ImportError("No module named libqtile.core._ffi_xcursors, be sure to run `./scripts/ffibuild`")
+    raise ImportError("No module named libqtile.backend.x11._ffi_xcursors, be sure to run `./scripts/ffibuild`")
 
 
 # Stolen from samurai-x
