+++ libqtile/scripts/qtile.py
@@ -27,7 +27,7 @@ from os import path, getenv
 
 from libqtile.log_utils import init_log, logger
 from libqtile import confreader
-from libqtile.core import xcore
+from libqtile.backend.x11 import xcore
 
 locale.setlocale(locale.LC_ALL, locale.getdefaultlocale())  # type: ignore
