+++ libqtile/widget/prompt.py
@@ -45,8 +45,8 @@ from libqtile.command_interface import CommandError, QtileCommandInterface
 from libqtile.log_utils import logger
 
 from . import base
-from .. import bar, hook, pangocffi, utils, xkeysyms
-from ..core import xcbq
+from libqtile import bar, hook, pangocffi, utils, xkeysyms
+from libqtile.backend.x11 import xcbq
 
 
 class AbstractCompleter(metaclass=abc.ABCMeta):
