+++ libqtile/widget/systray.py
@@ -26,9 +26,9 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from .. import bar, window
-from ..core import xcbq
 from . import base
+from libqtile import bar, window
+from libqtile.backend.x11 import xcbq
 
 import xcffib
 from xcffib.xproto import (ClientMessageEvent, ClientMessageData, EventMask,
