+++ scripts/ffibuild
@@ -2,4 +2,4 @@
 echo "building pango"
 python ./libqtile/pango_ffi_build.py
 echo "building xcursors"
-python ./libqtile/core/xcursors_ffi_build.py
+python ./libqtile/backend/x11/xcursors_ffi_build.py
