+++ setup.cfg
@@ -112,13 +112,13 @@ mypy_path = stubs
 python_version = 3.5
 [mypy-_cffi_backend]
 ignore_missing_imports = True
-[mypy-cairocffi._ffi]
+[mypy-cairocffi._generated.ffi]
 ignore_missing_imports = True
 [mypy-cairocffi.pixbuf]
 ignore_missing_imports = True
 [mypy-libqtile._ffi_pango]
 ignore_missing_imports = True
-[mypy-libqtile.core._ffi_xcursors]
+[mypy-libqtile.backend.x11._ffi_xcursors]
 ignore_missing_imports = True
 [mypy-xcffib._ffi]
 ignore_missing_imports = True
