+++ setup.py
@@ -68,6 +68,6 @@ setup(
     cmdclass={'install': CheckCairoXcb},
     cffi_modules=[
         'libqtile/pango_ffi_build.py:pango_ffi',
-        'libqtile/core/xcursors_ffi_build.py:xcursors_ffi',
+        'libqtile/backend/x11/xcursors_ffi_build.py:xcursors_ffi',
     ],
 )
