+++ test/test_config.py
@@ -25,7 +25,7 @@ import pytest
 
 from libqtile import confreader
 from libqtile import config, utils
-from libqtile.core import xcore
+from libqtile.backend.x11 import xcore
 
 tests_dir = os.path.dirname(os.path.realpath(__file__))
