+++ test/x11/test_xcore.py
@@ -1,0 +1,7 @@
+from libqtile.backend.x11 import xcore
+
+
+def test_keys(xephyr):
+    xc = xcore.XCore()
+    assert "a" in xc.get_keys()
+    assert "shift" in xc.get_modifiers()
